refactor: move reducer files to src/reducers/ folder
By clustering the reducer we set the foundation to further
refactoring the reducers by domain.
