+++ src/reducers/reducer.init.js
@@ -2,7 +2,7 @@ import { applyMiddleware, createStore } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 import thunk from 'redux-thunk';
 import { reducer } from './reducer';
-import rootSagas from './sagas';
+import rootSagas from '../sagas';
 
 const saga = createSagaMiddleware();
 const middlewares = [
